Crawlbot

Crawlbot

crawlbot :: String -> Crawlbot -> IO (Maybe Response)

crawlbotP :: String -> Maybe Proxy -> Crawlbot -> IO (Maybe Response)

crawlbotC :: String -> Command -> IO (Maybe Response)

data Crawlbot

Constructors

Crawlbot

Fields

crawlbotName :: String

Job name. This should be a unique identifier and can be used to modify your crawl or retrieve its output.

crawlbotSeeds :: Maybe [String]

Seed URL(s). By default Crawlbot will spider subdomains (e.g., a seed URL of http://www.diffbot.com will include URLs at http://blog.diffbot.com).

crawlbotApi :: Maybe Req

Diffbot API through which to process pages. E.g., Article to process matching links via the Article API.

crawlbotUrlCrawlLimit :: Maybe Limit

Limit crawled pages.

crawlbotUrlProcessLimit :: Maybe Limit

Limit processed pages.

crawlbotPageProcessPattern :: Maybe [String]

Specify strings to limit pages processed to those whose HTML contains any of the content strings.

crawlbotMaxToCrawl :: Maybe Int

Specify max pages to spider.

crawlbotMaxToProcess :: Maybe Int

Specify max pages to process through Diffbot APIs. Default: 10,000.

crawlbotRestrictDomain :: Maybe Bool

By default crawls will restrict to subdomains within the seed URL domain. Specify to 'Just False' to follow all links regardless of domain.

crawlbotNotifyEmail :: Maybe String

Send a message to this email address when the crawl hits the crawlbotMaxToCrawl or crawlbotMaxToProcess limit, or when the crawl completes.

crawlbotNotifyWebHook :: Maybe String

Pass a URL to be notified when the crawl hits the crawlbotMaxToCrawl or crawlbotMaxToProcess limit, or when the crawl completes. You will receive a POST with X-Crawl-Name and X-Crawl-Status in the headers, and the full JSON response in the POST body.

crawlbotDelay :: Maybe Double

Wait this many seconds between each URL crawled from a single IP address.

crawlbotRepeat :: Maybe Double

Specify the number of days as a floating-point (e.g. crawlbotRepeat=7.0) to repeat this crawl. By default crawls will not be repeated.

crawlbotOnlyProcessIfNew :: Maybe Bool

By default repeat crawls will only process new (previously unprocessed) pages. Set to 'Just False' to process all content on repeat crawls.

crawlbotMaxRounds :: Maybe Int

Specify the maximum number of crawl repeats. By default repeating crawls will continue indefinitely.

mkCrawlbot

Arguments

:: String
Job name.

-> Maybe [String]
Seed URLs.

-> Crawlbot

data Limit

Constructors

Pattern [String]
Specify strings to limit pages to those whose URLs contain any of the content strings. You can use the exclamation point to specify a negative string, e.g. "!product" to exclude URLs containing the string "product".

RegEx String
Specify a regular expression to limit pages to those URLs that match your expression.

data Response

Constructors

Response

Fields

responseString :: Maybe String

responseJobs :: Maybe [Job]

data Job

Constructors

Job

Fields

jobName :: String

jobType :: String

jobStatus :: JobStatus

jobSentDoneNotification :: Int

jobObjectsFound :: Int

jobUrlsHarvested :: Int

jobPageCrawlAttempts :: Int

jobPageCrawlSuccesses :: Int

jobPageProcessAttempts :: Int

jobPageProcessSuccesses :: Int

jobMaxRounds :: Int

jobRepeat :: Double

jobCrawlDelay :: Double

jobMaxToCrawl :: Int

jobMaxToProcess :: Int

jobObeyRobots :: Bool

jobRestrictDomain :: Bool

jobOnlyProcessIfNew :: Bool

jobSeeds :: [String]

jobRoundsCompleted :: Int

jobRoundStartTime :: UTCTime

jobCurrentTime :: UTCTime

jobApiUrl :: String

jobUrlCrawlPattern :: [String]

jobUrlProcessPattern :: [String]

jobPageProcessPattern :: [String]

jobUrlCrawlRegEx :: String

jobUrlProcessRegEx :: String

jobDownloadJson :: String

jobDownloadUrls :: String

jobNotifyEmail :: String

jobNotifyWebhook :: String

data JobStatus

Constructors

JobStatus

Fields

jobStatusCode :: Int

jobStatusMessage :: String

data Command

Constructors

Command

Fields

commandName :: String

Job name as defined when the crawl was created.

commandAction :: Action

Action to execute.

data Action

Constructors

Pause
Pause a crawl.

Resume
Resume a paused crawl.

Restart
Restart removes all crawled data while maintaining crawl settings.

Delete
Delete a crawl, and all associated data, completely.